feat(Composite): Merge the new learning path algorithm into the lib
Merge the new algorithm of finding learning paths into the lib project, and remove the old algorithm.
GreedyLearningPath is stopped/commented.
